+++ CHANGELOG.md
@@ -25,6 +25,8 @@
     (https://github.com/pulumi/pulumi-kubernetes/pull/639).
 -   Use dry-run support if available when diffing the actual and desired state of a resource
     (https://github.com/pulumi/pulumi-kubernetes/pull/649)
+-   Fix panic when `.metadata.label` is mistyped
+    (https://github.com/pulumi/pulumi-kubernetes/pull/655).
 
 ## 0.25.2 (July 11, 2019)